Adición columna size a la tabla domain_gserviceaccount

+++ aon-master/src/main/resources/com/code/aon/master/update/update.database.8.0.0.sql
@@ -3,7 +3,7 @@
 # Creation Date: 13/08/2014 
 
 
-ALTER TABLE `domain_gserviceaccount` ADD COLUMN `size` FLOAT NULL DEFAULT 0  AFTER `domain` ;
+ALTER TABLE `domain_gserviceaccount` ADD COLUMN `size` DOUBLE NULL DEFAULT 0  AFTER `domain` ;
 
 UPDATE `db_version` SET `version_number` = '8.0.1';
